Bandwidth limit should be float64.

+++ info/node.go
@@ -35,9 +35,9 @@ type Node struct {
 
 // Host storage publishable information
 type NodeStorage struct {
-	StoragePriceAsk   uint64 `sql:"storage_price_ask" json:"storage_price_ask"`
-	BandwidthPriceAsk uint64 `sql:"bandwidth_price_ask" json:"bandwidth_price_ask"`
-	StorageTimeMin    uint64 `sql:"storage_time_min" json:"storage_time_min"`
-	BandwidthLimit    uint64 `sql:"bandwidth_limit" json:"bandwidth_limit"`
-	CollateralStake   uint64 `sql:"collateral_stake" json:"collateral_stake"`
+	StoragePriceAsk   uint64  `sql:"storage_price_ask" json:"storage_price_ask"`
+	BandwidthPriceAsk uint64  `sql:"bandwidth_price_ask" json:"bandwidth_price_ask"`
+	StorageTimeMin    uint64  `sql:"storage_time_min" json:"storage_time_min"`
+	BandwidthLimit    float64 `sql:"bandwidth_limit" json:"bandwidth_limit"`
+	CollateralStake   uint64  `sql:"collateral_stake" json:"collateral_stake"`
 }
